Create Time Series Case Study
In this project, we forecast the sales and the demand for the next 6 months, that would help you manage the revenue and inventory accordingly. The store caters to 7 different market segments and in 3 major categories. We want to forecast at this granular level, so you subset your data into 21 (7*3) buckets before analysing these data. From these 21 buckets, we identify the 5 most profitable market segments , byidentifying the segment, with the lowest coefficient of variation